requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
exceptiongroup==1.3.0
google-auth==2.40.3
google-genai==1.33.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1

# untuk image segmentation
torch==2.4.1
torchvision==0.19.1
opencv-python
numpy
pillow
# LangSAM (Language Segment Anything) langsung dari GitHub
git+https://github.com/luca-medeiros/lang-segment-anything.git


